사용자가 입력할 수 있는 리뷰페이지를 만들어라 사용자가 입력하는 데이터를 서버에 저장할 수 있도록 리뷰페이지를 만들고 있습니다. 그 밖에 PostReviewPage => PostReview로 컴포넌트명 변경하였고, 모바일화면 사이드바 햄버거메뉴를 클릭했을 때 배경화면이 어두워지도록 자료형을 문자열('true')에서 boolean(true)값으로 변경했습니다.
